Solution of a Programming Problem
Write a C++ program that declares an array alpha of 50 components of type double. Initialize the array so that the first 25 components are equal to the cube of the index variable, and the last 25 components are equal to four times the index variable. Output the array so that 10 elements per line are printed.
Use pointers notation to solve this question.